Initial commit of TyCode
Tycode is an AI coding assistant that runs in either as a VSCode extension or as
a standalone CLI application. I originally built this because I wasn't able to
customize existing tools enough. What is in this commit is far from finished but
this is the first version that is more or less usable.

All of the code is AI slop. I've attempted to guide the architecture to make it
possible to edit, but huge portions of the codebase (mainly the vscode
typescript) I have never even reviewed.