Improve startup experience
Previously the experience starting tycode was you would default to bedrock as
your provider and grok as the default model. Bedrock doesn't support grok so
any request would immediately error that "grok is not supported in bedrock".

Now, you can configure a cost preference and tycode will use the best model
the current provider supports within the configured price range. Additionally,
we do not default to Bedrock, we default to "Mock" in error every request mode
to give better feedback that a provider needs to be configured. Finally, if you
don't have a provider or price configuration set you get a warning/error.