Add a new benchmark that attempts to measure modify_file success
This change adds a new benchmark that presents a broken a complex rust file.
Models are challenged to modify the file until it compiles (and encouraged to
not just comment everything out). The intent is to measure tools accuracy with
the modify_file tool (which currently uses search/replace). I suspect that GPT5
is not wortking well with search/replace because its so heavily biased towards
the GPT diff format. The idea here is I need a base line to measure new diff
tools against.

I also added the GPT diff format as a tool, but currently nothing is using it.
There are a few minor tweeks to attempt to make modify_file more successful.