Significant improvements to file editing
This change removes the single retry without providing feedback. I think that
was counter productive and wasting tokens; generally the feedback was quite
small and it was relatively rare a retry with no feedback would produce a
better result. Along with that I fixed some bugs where tools where not getting
marked as failed when a model made an invalid request.

Second, this change allows partial tool calls to succeed. For example, if a
model made 3 modify_file requests and 1 of 3 failed, we accept the 2 correct
patches and reject the invalid patch.

Third, we now reject tool calls of different "category". For example, models
would be too optimistic and make a modify_file tool call along with a
complete_task call. The modify_file could be a completely insane diff but the
model would also complete_task. Now we accept the least impactful tool call
category and drop others,

Finally, this change makes a bunch of improvements to the apply_codex_patch
tool (which is still not enabled). In benchmarks GPT-5 does significantly
better with the path tool than with the search/replace tool.