Add support for settings profiles
You can now have multiple settings files and provide a "profile" which will
resolve to settings_{profile}.toml. You can save settings as a new profile or
switch profiles with slash commands or provide a profile on startup to the CLI.
This makes it easy to swap between settings - you may have a low cost profile
and a claude code profile and swap between them depending on what you are doing.