Improvements to coordinator mode
This change makes various improvements to coordinator mode. Coordinator now has
a dedicated tool to spawn the coder - it was occasionally getting mixed up and
spawning the default agent or the agent for benchmarking. This change combines
propose task list and update task list to just 1 tool that sets the task list
and gives strong guidance (which is mostly ignored) to combine updating the task
list with other meaningful work. Error handling and turn ending logic is
modified to accomodate models incorrectly updating the task list as a stand
alone tool call.

This change also adds task level review mode. If the coordinator spawns a task
it will push 2 agents to the stack a review agent and a work agent. The review
agent can approve or rejecct the work.