Attempt to handle large context more gracefully
Typically, people run out of context when opening tycode in a huge directory and
the auto-context attempts to put all files that exist in the context window.
This change now makes auto-context not contain a directory listing and gives the
user a warning when there are too many files in the current directory.

This change also adds automatic compaction for if conversation history gets too
long. I rarely encounter this so its probably not too useful, but now its here.