Improve code review agent to reduce token usage & improve quality
The code reviewer was getting pushed on the agent stack before the coding agent
so the coding agent would pop to the review agent which is pretty natural,
however this had 2 problems: 1/ the reviewer couldnt see what the coder did 2/
the reviewers feedback went to the coordinator (which had to use tokens to spawn
the coder again). Now we push a code reviewer when the coder complete_task and
if the review is approved pop up twice.