Invoke git to determine ignored files
This change fixes bugs related to the previous manually implement gitignore
filtering. We now use git to process the gitignore list and avoids a bunch of
bugs in our custom implementation. This also fixes a bug where .gitignore in
subdirectories was not applied. If you don't know how to use multi workspaces
and instead just open a parent directory, this now respects .gitignore