Improve file context building
Previously if you opened tycode in too large of a directory it would take
forever to explore the files that exist (each invoked git) and if the file list
was > the auto context limit we would include 0 files in context. This
effectively would break tycode since the agents cannot see file lists so we sent
a warning that you should setup git. However, this was all a rough getting
started experience.

This change 1/ uses the ignore crate which is significantly faster than invoking
git 2/ uses BFS and always puts some files in the context even if there are too
many files to fit the entire file tree in the context 3/ compacts the file list
a bit better.