Add a hack to improve bedrock prompt caching
Prompt caching doesn't work well with the rust SDK due to random hashmap
ordering. I cannot find any way to hack in deterministic sorting in to the rust
SDK or smithy-rs. The best fix I can come up with is retry creating the hashmap
over and over until we happen upon sorted order. This might be the only
practical usecase of bogo sort.