Introduce TypeAnalyzer for code intelligence
This change introduces a new "TypeAnalyzer" which defines an interface for code
intelligence. This currently has 3 methods: search by name (to find a FQN), get
a rust outline/docs for a type, and get build status. I've built two
implementations for rust - a syn based and LSP based implementation. The lsp
implementation was turning out to be impossible to generalize so the syn
implementation is just a lot simpler.