Add XML and Json fallback tool parsing
Most models tend to have a better time generating xml responses rather than
json (and it seems like anthropics and xai just convert xml in to json to
avoid having to train the model to reliably generate json). Additionally, if you
disable all built in tools in claude code it seems like it is not possible for
the model to output tool calls (and we want to disable the built in tools - it
saves about 10k extra tokens per request and avoids confusing the model about
which tools it has)

This change parses text blocks for tool calls in xml and json and if they appear
to be tool blocks, parses them out from the text message. Additionally, we can
have the system prompt either present tools in openai json schema or just as
additionally instructions in the system prompt. We treat this as a preference,
as a design goal, we'll take anything that looks like a tool call no matter what
mode we are in to try and take what the models give us. We have a setting which can
force a preference for xml tool calls. We also now allow providers to express
preference via a "Tweaks" construct.

All in all, this should 1/ fix a bug where older versions of claude code cannot
make tool calls (since we disabled all of the tool calls except for 1 introduced
in a newer version of claude code) 2/ save about 10k tokens on the system prompt
in claude code 3/ allow weaker models that tend to make XML tool calls to work
better in tycode.