Remove the benchmark
Having the benchmark coupled to the internal structure of tycode is not
required. We can have a generic benchmark runner that would run any CLI agent.
I've had a fork of this locally that runs SWE bench which has been annoying to
maintain the the mainline version does basically nothing so its not worth
having. I'm moving this in to a benchmark runner in its own repo with no tie to
the internals of tycode.