Introduce memory management
This change introduces the start of the memory system. The intent is that, over
time, tycode will learn from your past interactions and (ideally) remember
feedback and improve over time. This change does not introduce any functional
changes - memories are stored but are never read or injected in to the system
prompt.

My thoughts for a memory architecture are that I'm unlikely to get it right and
will likely need to iterate. So, I want to start with having an agent that can
extract learnings and simply just store them in a log. This will allow us to
build various views over the raw data and experiment with different ways to make
memories useful. Memories are off by default. Opt in through settings.