Introduce "Autonomy Level"
I initially conflated the concept of autonomy with sercurity and agents. The
idea was you would pick agents or security models that match the level of
autonomy you want the agent to have. This isn't particularly intuitive and
wasn't implemented that well.

My new idea is we have: "autonomy" (i.e. how much oversight you prefer to have
over what the model is doing). I plan to remove the security settings or
implement it differently to make setup more intuitive.