Change vscode UI to swap settings profiles rather than providers
The botton of the chat has a drop down menu to force the use of a particular AI
provider. This was a super old setting from before we had settings or settings
profiles. Now swapping settings profiles makes more sense as that allows you
to control everything (including the provider).