Delete security settings
This was an incoherent idea. You need security mode = all for the agent to do
anything useful (i.e. run commands). Users were stumbling over not setting it
and having the model being dumb. There was no way to allow or block specific
commands so you basically had to set it to all. I'm just deleting it. I've
slightly improved the state of the world by blocking 'rm' at least which was not
blocked in security mode = all.